+++ lib/fluent/tranquility/plugin.rb
@@ -13,6 +13,13 @@ module Fluent
       desc 'Druid Dataset'
       config_param :dataset, :string
 
+      config_section :retries do
+        config_param :max, :integer, default: 5
+        config_param :interval, :integer, default: 1
+        config_param :interval_randomness, :integer, default: 0.5
+        config_param :backoff_factor, :integer, default: 2
+      end
+
       def start
         super
         @writer = Writer.new.tap do |w|